+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `Grid2` is now a subclass of `Rect`
+- `Grid2.__str__()` now includes step and origin values instead of just the rectangle coordiantes
 - `util.snap_num()` parameter `mult` is now typed as `float`
 - `util.snap_num()` return type is now `float`
 - `util.snap_num()` parameter `snap_fn` is now optional (defaults to the built-in `round` function)